# demo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "email": "yinjie_luo@intsig.net"
        }
    },
    "host": "http://192.168.6.50:10006",
    "basePath": "/edapi/bankbills/",
    "paths": {
...
        "/edapi/bankbills/analyse/jyhz/abnormal/transaction/high_frequency": {
            "get": {
                "description": "交易汇总-异常交易-高频交易",
                "tags": [
                    "交易汇总"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.AbnormalTransactionHighFrequencyResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/jyhz/abnormal/transaction/large": {
            "get": {
                "description": "交易汇总-异常交易-大额交易",
                "tags": [
                    "交易汇总"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.AbnormalTransactionLargeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/jyhz/abnormal/transaction/suspicious": {
            "get": {
                "description": "交易汇总-异常交易-可疑交易",
                "tags": [
                    "交易汇总"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.AbnormalTransactionSuspiciousResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/jyhz/balance/fluctuation": {
            "get": {
                "description": "交易汇总-余额波动",
                "tags": [
                    "交易汇总"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.BalanceFluctuationResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/jyhz/transaction/distribution": {
            "get": {
                "description": "交易汇总-交易分布",
                "tags": [
                    "交易汇总"
                ],
                "parameters": [
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "交易分布展示 0:交易金额 1:交易笔数 默认展示交易金额页",
                        "name": "disp_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.TransactionDistribution"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/lsyz/authenticity": {
            "get": {
                "description": "流水验证-数据真实性",
                "tags": [
                    "流水验证"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.LsyzAResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rest.AbnormalTransactionHighFrequencyResp": {
            "type": "object",
            "properties": {
                "high_frequency_trading": {
                    "description": "高频交易",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.HighFrequencyTradingI"
                    }
                }
            }
        },
        "rest.AbnormalTransactionLargeResp": {
            "type": "object",
            "properties": {
                "large_num": {
                    "description": "大额交易流水总数",
                    "type": "integer"
                },
                "large_transaction": {
                    "description": "大额交易",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.LargeTransactionI"
                    }
                }
            }
        },
        "rest.AbnormalTransactionSuspiciousResp": {
            "type": "object",
            "properties": {
                "suspicious_num": {
                    "description": "可以交易流水总数",
                    "type": "integer"
                },
                "suspicious_transaction": {
                    "description": "可疑交易",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.SuspiciousTransactionI"
                    }
                }
            }
        },
...
        "rest.BalanceFluctuationResp": {
            "type": "object",
            "properties": {
                "balance_fluctuation": {
                    "description": "余额波动",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.BalanceInfo"
                    }
                }
            }
        },
        "rest.BalanceInfo": {
            "type": "object",
            "properties": {
                "avg_balance": {
                    "description": "月均余额",
                    "type": "integer"
                },
                "timer": {
                    "description": "月 2022-01",
                    "type": "string"
                },
                "volatility": {
                    "description": "波动率 标注差 20",
                    "type": "integer"
                }
            }
        },
...
        "rest.HighFrequencyTradingI": {
            "type": "object",
            "properties": {
                "income": {
                    "description": "收入",
                    "type": "integer"
                },
                "income_num": {
                    "description": "收入笔数",
                    "type": "integer"
                },
                "name": {
                    "description": "对手方",
                    "type": "string"
                },
                "pay": {
                    "description": "支出",
                    "type": "integer"
                },
                "pay_num": {
                    "description": "支出笔数",
                    "type": "integer"
                }
            }
        },
        "rest.InterestReportI": {
            "type": "object",
            "properties": {
                "estimate_interest": {
                    "description": "预估利息 单位元",
                    "type": "integer"
                },
                "interest": {
                    "description": "利息 单位元",
                    "type": "integer"
                },
                "interest_time": {
                    "description": "结息日期 后端格式化之后返回",
                    "type": "string"
                },
                "summary": {
                    "description": "交易摘要",
                    "type": "string"
                },
                "trade_time": {
                    "description": "交易日期 后端格式化之后返回",
                    "type": "string"
                }
            }
        },
...
        "rest.LargeTransactionI": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "amount": {
                    "description": "金额",
                    "type": "integer"
                },
                "name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "proportion": {
                    "description": "当月占比",
                    "type": "string"
                },
                "time": {
                    "description": "时间",
                    "type": "string"
                }
            }
        },
        "rest.LsyzAResp": {
            "type": "object",
            "properties": {
                "abnormal_times": {
                    "description": "结息金额异常次数",
                    "type": "integer"
                },
                "actual_interest": {
                    "description": "实结息",
                    "type": "integer"
                },
                "end_date": {
                    "description": "筛选之后显示的流水结束时间",
                    "type": "string"
                },
                "interest_report": {
                    "description": "结息报表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.InterestReportI"
                    }
                },
                "should_interest": {
                    "description": "应结息",
                    "type": "integer"
                },
                "start_date": {
                    "description": "筛选之后显示的流水开始时间",
                    "type": "string"
                },
                "total_interest": {
                    "description": "总结息金额",
                    "type": "integer"
                }
            }
        },
...
        "rest.Result1": {
            "type": "object",
            "properties": {
                "data": {},
                "err": {},
                "ret": {
                    "type": "integer"
                }
            }
        },
...
        "rest.SuspiciousTransactionI": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "amount": {
                    "description": "金额",
                    "type": "integer"
                },
                "name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "time": {
                    "description": "时间",
                    "type": "string"
                }
            }
        },
...
        "rest.TradeAmountI": {
            "type": "object",
            "properties": {
                "income": {
                    "description": "当月收入或者收入笔数",
                    "type": "integer"
                },
                "pay": {
                    "description": "当月支出或者支出笔数",
                    "type": "integer"
                },
                "timer": {
                    "description": "月 2022-01",
                    "type": "string"
                },
                "total": {
                    "description": "当月总流水或者交易总笔数",
                    "type": "integer"
                }
            }
        },
        "rest.TransactionDistribution": {
            "type": "object",
            "properties": {
                "trade_data": {
                    "description": "交易金额或者交易笔数信息 根据disp_type来填充返回值",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.TradeAmountI"
                    }
                }
            }
        }
    }
}